+++ Stitch/Graph/Node/Layer/ViewModel/LayerNodeViewModel.swift
@@ -14,338 +14,6 @@ import StitchEngine
 typealias LayerNode = NodeViewModel
 typealias LayerNodes = [LayerNode]
 
-final class LayerInputUnpackedPortObserver {
-    let layerPort: LayerInputPort
-    let layer: Layer
-    
-    var port0: InputLayerNodeRowData
-    var port1: InputLayerNodeRowData
-    var port2: InputLayerNodeRowData
-    
-    init(layerPort: LayerInputPort,
-         layer: Layer,
-         port0: InputLayerNodeRowData,
-         port1: InputLayerNodeRowData,
-         port2: InputLayerNodeRowData) {
-        self.layerPort = layerPort
-        self.layer = layer
-        self.port0 = port0
-        self.port1 = port1
-        self.port2 = port2
-    }
-    
-    /// Only to be used by `allPorts` helper.
-    private var _allAvailablePorts: [InputLayerNodeRowData] {
-        [port0, port1, port2]
-    }
-}
-
-extension LayerInputUnpackedPortObserver {
-    @MainActor
-    func getParentPortValuesList() -> PortValues {
-        // TODO: ignore unpacked ports until we support unpacking
-        if !FeatureFlags.SUPPORTS_LAYER_UNPACK {
-            return []
-        }
-        
-        let allRawValues: PortValuesList = allPorts.map { $0.allLoopedValues }
-        let lengthenedValues: PortValuesList = allRawValues.lengthenArrays()
-        
-        // Remap values so we can process packing logic
-        let remappedValues = lengthenedValues.remapValuesByLoop()
-        let packedValues = remappedValues.map { valuesList in
-            self.layerPort.packValues(from: valuesList, 
-                                      layer: self.layer)
-        }
-        
-        return packedValues
-    }
-    
-    @MainActor
-    var allPorts: [InputLayerNodeRowData] {        
-        // TODO: ignore unpacked ports until we support unpacking
-        if !FeatureFlags.SUPPORTS_LAYER_UNPACK {
-            return []
-        }
-
-        guard let portsToUse = layerPort.unpackedPortCount(layer: self.layer) else {
-            return []
-        }
-        
-        let relevantPorts = self._allAvailablePorts.prefix(upTo: portsToUse)
-        assertInDebug(portsToUse == relevantPorts.count)
-        
-        return Array(relevantPorts)
-    }
-    
-    @MainActor
-    func createSchema() -> [LayerInputDataEntity] {
-        self.allPorts.map { $0.createSchema() }
-    }
-    
-    @MainActor
-    func initializeDelegate(_ node: NodeDelegate) {
-        self.allPorts.forEach {
-            $0.initializeDelegate(node)
-        }
-    }
-    
-    @MainActor
-    /// From packed values, unpacks them for unpack layer input scenario.
-    func updateValues(from packedValues: PortValues,
-                      layerNode: LayerNodeViewModel) {
-        let unpackedValues = packedValues.map { self.layerPort.unpackValues(from: $0) }
-        
-        guard let unpackedPortCount = unpackedValues.first??.count else {
-            fatalErrorIfDebug()
-            return
-        }
-        
-        // Remap values to be all organized for a particular port
-        (0..<unpackedPortCount).forEach { portId in
-            guard let unpackedId = UnpackedPortType(rawValue: portId) else {
-                fatalErrorIfDebug()
-                return
-            }
-            
-            // Grab loop of values from unpacked array for this indexed unpacked port
-            let values = unpackedValues.map {
-                guard let value = $0?[safe: portId] else {
-                    fatalErrorIfDebug()
-                    return PortValue.none
-                }
-                
-                return value
-            }
-            
-            let portTypeId: LayerInputKeyPathType = .unpacked(unpackedId)
-            let layerId = LayerInputType(layerInput: self.layerPort,
-                                        portType: portTypeId)
-            let rowObserver = layerNode[keyPath: layerId.layerNodeKeyPath].rowObserver
-            
-            // Update row observer values per usual
-            rowObserver.updateValues(values)
-        }
-    }
-}
-
-///// Needs to be class for StitchEngine which assumes reference objects with its mutation logic
-//final class LayerInputObserver {
-//    var id: NodeIOCoordinate    // ID used for NodeRowCalculatable
-//    var type: LayerInputMode
-//    
-//    init(id: NodeIOCoordinate,
-//         type: LayerInputObserverMode) {
-//        self.id = id
-//        self.type = type
-//    }
-//}
-
-// TODO: move
-enum LayerInputMode {
-    case packed
-    case unpacked
-}
-
-// Must be a class for coordinate keypaths, which expect a reference type on the other end.
-@Observable
-final class LayerInputObserver {
-    // Not intended to be used as an API given both data payloads always exist
-    // Variables here necessary to ensure keypaths logic works
-    var _packedData: InputLayerNodeRowData
-    var _unpackedData: LayerInputUnpackedPortObserver
-    
-    let layer: Layer
-    var port: LayerInputPort
-    
-    @MainActor
-    init(from schema: LayerNodeEntity, port: LayerInputPort) {
-        self.layer = schema.layer
-        self.port = port
-        
-        self._packedData = .empty(.init(layerInput: port,
-                                        portType: .packed),
-                                  layer: schema.layer)
-        
-        self._unpackedData = .init(layerPort: port,
-                                   layer: schema.layer,
-                                   port0: .empty(.init(layerInput: port,
-                                                       portType: .unpacked(.port0)),
-                                                 layer: schema.layer),
-                                   port1: .empty(.init(layerInput: port,
-                                                       portType: .unpacked(.port1)),
-                                                 layer: schema.layer),
-                                   port2: .empty(.init(layerInput: port,
-                                                       portType: .unpacked(.port2)),
-                                                 layer: schema.layer))
-    }
-}
-
-enum LayerInputObserverMode {
-    case packed(InputLayerNodeRowData)
-    case unpacked(LayerInputUnpackedPortObserver)
-}
-
-extension LayerInputObserver {
-    @MainActor
-    var mode: LayerInputMode {
-        if self._unpackedData.allPorts.contains(where: { $0.canvasObserver.isDefined }) {
-            return .unpacked
-        }
-        
-        return .packed
-    }
-    
-    /// Updates all-up values, handling scenarios like unpacked if applicable.
-    @MainActor func updatePortValues(_ values: PortValues) {
-        // Updating the packed observer will always update unpacked observers if the mode is set as unpacked
-        self._packedData.rowObserver.updateValues(values)
-    }
-    
-    /// All-up values for this port
-    var allLoopedValues: PortValues {
-        self._packedData.allLoopedValues
-    }
-    
-    @MainActor
-    var observerMode: LayerInputObserverMode {
-        switch self.mode {
-        case .packed:
-            return .packed(self._packedData)
-        case .unpacked:
-            return .unpacked(self._unpackedData)
-        }
-    }
-    
-    @MainActor
-    var values: PortValues {
-        switch self.mode {
-        case .packed:
-            return self._packedData.rowObserver.values
-        case .unpacked:
-            return self._unpackedData.getParentPortValuesList()
-        }
-    }
-    
-    var graphDelegate: GraphDelegate? {
-        // Hacky solution, just get row observer delegate from packed data
-        self._packedData.rowObserver.nodeDelegate?.graphDelegate
-    }
-    
-    @MainActor var activeValue: PortValue {
-        let activeIndex = self.graphDelegate?.activeIndex ?? .init(.zero)
-        let values = self.values
-        
-        guard let value = values[safe: activeIndex.adjustedIndex(values.count)] else {
-            fatalErrorIfDebug()
-            return values.first ?? .none
-        }
-        
-        return value
-    }
-    
-    @MainActor
-    var allInputData: [InputLayerNodeRowData] {
-        switch self.observerMode {
-        case .packed(let packedData):
-            return [packedData]
-        case .unpacked(let unpackedObserver):
-            return unpackedObserver.allPorts
-        }
-    }
-    
-    @MainActor func initializeDelegate(_ node: NodeDelegate) {
-        switch self.mode {
-        case .packed:
-            self._packedData.initializeDelegate(node)
-        case .unpacked:
-            self._unpackedData.initializeDelegate(node)
-        }
-    }
-    
-    @MainActor func getAllCanvasObservers() -> [CanvasItemViewModel] {
-        switch self.observerMode {
-        case .packed(let packedData):
-            if let canvas = packedData.canvasObserver {
-                return [canvas]
-            }
-            return []
-        case .unpacked(let unpackedData):
-            return unpackedData.allPorts.compactMap {
-                $0.canvasObserver
-            }
-        }
-    }
-    
-    @MainActor func toggleMode() {
-        let nodeId = self._packedData.rowObserver.id.nodeId
-        let parentGroupNodeId = self.graphDelegate?.groupNodeFocused
-        
-        guard let node = self.graphDelegate?.getNodeViewModel(nodeId),
-              let layerNode = node.layerNode else {
-            fatalErrorIfDebug()
-            return
-        }
-        
-        switch self.mode {
-        case .packed:
-            // Reset packed state
-            self._packedData.resetOnPackModeToggle()
-            
-            self._unpackedData.allPorts.forEach { unpackedPort in
-                var unpackSchema = unpackedPort.createSchema()
-                unpackSchema.canvasItem = .init(position: .zero,
-                                                zIndex: .zero,
-                                                parentGroupNodeId: parentGroupNodeId)
-                unpackedPort.update(from: unpackSchema,
-                                    layerInputType: unpackedPort.id,
-                                    layerNode: layerNode,
-                                    nodeId: nodeId,
-                                    nodeDelegate: node)
-            }
-            
-        case .unpacked:
-            guard let packedKeyPath = self.__packedData.rowObserver.id.keyPath else {
-                fatalErrorIfDebug()
-                return
-            }
-
-            // Reset unpacked state
-            self._unpackedData.allPorts.forEach {
-                $0.resetOnPackModeToggle()
-            }
-            
-            var packedSchema = self._packedData.createSchema()
-            packedSchema.canvasItem = .init(position: .zero,
-                                            zIndex: .zero,
-                                            parentGroupNodeId: parentGroupNodeId)
-            
-            self._packedData.update(from: packedSchema,
-                                    layerInputType: packedKeyPath,
-                                    layerNode: layerNode,
-                                    nodeId: nodeId,
-                                    nodeDelegate: node)
-        }
-        
-        self.graphDelegate?.updateGraphData(document: nil)
-    }
-    
-    /// Helper only intended for use with ports that don't support unpacked mode.
-    @MainActor
-    var rowObserver: InputNodeRowObserver {
-        assertInDebug(self.mode == .packed)
-        return self._packedData.rowObserver
-    }
-}
-
-extension InputLayerNodeRowData {
-    /// Resets canvas data and connections when toggled between pack/unpack state.
-    func resetOnPackModeToggle() {
-        self.rowObserver.upstreamOutputCoordinate = nil
-        self.canvasObserver = nil
-    }
-}
-
 // primary = hidden via direct click from user
 // secondary = hidden because was a child of a group that was primary-hidden
 // none = not hidden at all
