+++ Stitch/Graph/Node/Port/ViewModel/Layer/LayerInputObserver.swift
@@ -1,0 +1,213 @@
+//
+//  LayerInputObserver.swift
+//  Stitch
+//
+//  Created by Elliot Boschwitz on 8/14/24.
+//
+
+import Foundation
+import StitchSchemaKit
+
+// Must be a class for coordinate keypaths, which expect a reference type on the other end.
+@Observable
+final class LayerInputObserver {
+    // Not intended to be used as an API given both data payloads always exist
+    // Variables here necessary to ensure keypaths logic works
+    var _packedData: InputLayerNodeRowData
+    var _unpackedData: LayerInputUnpackedPortObserver
+    
+    let layer: Layer
+    var port: LayerInputPort
+    
+    @MainActor
+    init(from schema: LayerNodeEntity, port: LayerInputPort) {
+        self.layer = schema.layer
+        self.port = port
+        
+        self._packedData = .empty(.init(layerInput: port,
+                                        portType: .packed),
+                                  layer: schema.layer)
+        
+        self._unpackedData = .init(layerPort: port,
+                                   layer: schema.layer,
+                                   port0: .empty(.init(layerInput: port,
+                                                       portType: .unpacked(.port0)),
+                                                 layer: schema.layer),
+                                   port1: .empty(.init(layerInput: port,
+                                                       portType: .unpacked(.port1)),
+                                                 layer: schema.layer),
+                                   port2: .empty(.init(layerInput: port,
+                                                       portType: .unpacked(.port2)),
+                                                 layer: schema.layer))
+    }
+}
+
+extension LayerInputObserver {
+    @MainActor
+    var mode: LayerInputMode {
+        if self._unpackedData.allPorts.contains(where: { $0.canvasObserver.isDefined }) {
+            return .unpacked
+        }
+        
+        return .packed
+    }
+    
+    /// Updates all-up values, handling scenarios like unpacked if applicable.
+    @MainActor func updatePortValues(_ values: PortValues) {
+        // Updating the packed observer will always update unpacked observers if the mode is set as unpacked
+        self._packedData.rowObserver.updateValues(values)
+    }
+    
+    /// All-up values for this port
+    var allLoopedValues: PortValues {
+        self._packedData.allLoopedValues
+    }
+    
+    @MainActor
+    var observerMode: LayerInputObserverMode {
+        switch self.mode {
+        case .packed:
+            return .packed(self._packedData)
+        case .unpacked:
+            return .unpacked(self._unpackedData)
+        }
+    }
+    
+    @MainActor
+    var values: PortValues {
+        switch self.mode {
+        case .packed:
+            return self._packedData.rowObserver.values
+        case .unpacked:
+            return self._unpackedData.getParentPortValuesList()
+        }
+    }
+    
+    var graphDelegate: GraphDelegate? {
+        // Hacky solution, just get row observer delegate from packed data
+        self._packedData.rowObserver.nodeDelegate?.graphDelegate
+    }
+    
+    @MainActor var activeValue: PortValue {
+        let activeIndex = self.graphDelegate?.activeIndex ?? .init(.zero)
+        let values = self.values
+        
+        guard let value = values[safe: activeIndex.adjustedIndex(values.count)] else {
+            fatalErrorIfDebug()
+            return values.first ?? .none
+        }
+        
+        return value
+    }
+    
+    @MainActor
+    var allInputData: [InputLayerNodeRowData] {
+        switch self.observerMode {
+        case .packed(let packedData):
+            return [packedData]
+        case .unpacked(let unpackedObserver):
+            return unpackedObserver.allPorts
+        }
+    }
+    
+    @MainActor func initializeDelegate(_ node: NodeDelegate) {
+        switch self.mode {
+        case .packed:
+            self._packedData.initializeDelegate(node)
+        case .unpacked:
+            self._unpackedData.initializeDelegate(node)
+        }
+    }
+    
+    @MainActor func getAllCanvasObservers() -> [CanvasItemViewModel] {
+        switch self.observerMode {
+        case .packed(let packedData):
+            if let canvas = packedData.canvasObserver {
+                return [canvas]
+            }
+            return []
+        case .unpacked(let unpackedData):
+            return unpackedData.allPorts.compactMap {
+                $0.canvasObserver
+            }
+        }
+    }
+    
+    @MainActor func toggleMode() {
+        let nodeId = self._packedData.rowObserver.id.nodeId
+        let parentGroupNodeId = self.graphDelegate?.groupNodeFocused
+        
+        guard let node = self.graphDelegate?.getNodeViewModel(nodeId),
+              let layerNode = node.layerNode else {
+            fatalErrorIfDebug()
+            return
+        }
+        
+        switch self.mode {
+        case .packed:
+            // Reset packed state
+            self._packedData.resetOnPackModeToggle()
+            
+            self._unpackedData.allPorts.forEach { unpackedPort in
+                var unpackSchema = unpackedPort.createSchema()
+                unpackSchema.canvasItem = .init(position: .zero,
+                                                zIndex: .zero,
+                                                parentGroupNodeId: parentGroupNodeId)
+                unpackedPort.update(from: unpackSchema,
+                                    layerInputType: unpackedPort.id,
+                                    layerNode: layerNode,
+                                    nodeId: nodeId,
+                                    nodeDelegate: node)
+            }
+            
+        case .unpacked:
+            guard let packedKeyPath = self.__packedData.rowObserver.id.keyPath else {
+                fatalErrorIfDebug()
+                return
+            }
+
+            // Reset unpacked state
+            self._unpackedData.allPorts.forEach {
+                $0.resetOnPackModeToggle()
+            }
+            
+            var packedSchema = self._packedData.createSchema()
+            packedSchema.canvasItem = .init(position: .zero,
+                                            zIndex: .zero,
+                                            parentGroupNodeId: parentGroupNodeId)
+            
+            self._packedData.update(from: packedSchema,
+                                    layerInputType: packedKeyPath,
+                                    layerNode: layerNode,
+                                    nodeId: nodeId,
+                                    nodeDelegate: node)
+        }
+        
+        self.graphDelegate?.updateGraphData(document: nil)
+    }
+    
+    /// Helper only intended for use with ports that don't support unpacked mode.
+    @MainActor
+    var rowObserver: InputNodeRowObserver {
+        assertInDebug(self.mode == .packed)
+        return self._packedData.rowObserver
+    }
+}
+
+extension InputLayerNodeRowData {
+    /// Resets canvas data and connections when toggled between pack/unpack state.
+    func resetOnPackModeToggle() {
+        self.rowObserver.upstreamOutputCoordinate = nil
+        self.canvasObserver = nil
+    }
+}
+
+enum LayerInputMode {
+    case packed
+    case unpacked
+}
+
+enum LayerInputObserverMode {
+    case packed(InputLayerNodeRowData)
+    case unpacked(LayerInputUnpackedPortObserver)
+}
